Add timestamp to infinite embed
